move inspect_dump and get_devnet_addresses into evm_bots
